requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
polars==0.20.31
xarray==2024.3.0
netCDF4==1.7.1
geopandas==0.14.4
httpx==0.27.0
google-auth==2.29.0
google-api-python-client==2.128.0
python-dotenv==1.0.1
pydantic==2.7.1